Format error messages consistently
The error messages are now presented in this format:

    JIRA::REST Error[<CODE> - <TEXT>]:
    <CONTENT>

<CODE> is the HTTP response code. If the error isn't related to a REST
HTTP invokation, then we use 500.

<TEXT> is the textual representation of <CODE>, which we get using the
HTTP::Status::status_message routine.

<CONTENT> is whatever content comes with the HTTP response. JIRA errors
usually come in a JSON structure that we parse to present only the
textual message. Other Content-Types are parsed as well: text/plain and
text/html (with HTML::TreeBuilder). Otherwise, the user gets a message
telling him which unsupported Content-Type we got from the HTTP
response.