Do not prepend default API prefix to absolute URLs
This change is intended to allow one to use directly the absolute URLs
returned by JIRA on the JSON resources to invoke JIRA::REST::GET on
them. However, it doesn't work yet because we need that the issue [0]
be fixed on REST::Client first. Anyway, as it doesn't currently work,
at least when that issue is fixed, JIRA::REST will be immediately
usable for that.

Ref0: https://rt.cpan.org/Ticket/Display.html?id=119790