Do not send Authorization header during sessions
When we pass the Authorization header in a HTTP call Jira increases the
login count for the authenticated user.

If we acquire a session cookie (passing the 'session' option to the
constructor) we don't need to pass the Authorization header for
subsequent calls, because Jira uses the cookie as an authentication
token already.

So, now we refrain to pass the Authorization header when we are in a
session.