Adds the 'session' option to the constructor
In retrospect, the new_session constructor was a bad idea. A new boolean
option called 'session' to the default constructor makes the code
simpler without imposing a noticeable burden to the user.

This script implements this new option and refactors the new_session
constructor so that it can take advantage of it. The new_session
constructor documentation says that it may be deprecated in the future.